refactor huge hooks
