+++ src/components/screens/SwapScreen/hooks/index.ts
@@ -1,3 +1,2 @@
 // SwapScreen-specific hooks - simplified and professional
-export { useTransactionStatus } from "./useTransactionStatus";
 export { useSwapPathFinding } from "./useSwapPathFinding";
