+++ src/i18n/locales/en/translations.json
@@ -357,7 +357,8 @@
     },
     "errors": {
       "failedToPrepareTransaction": "Failed to prepare swap transaction",
-      "swapTransactionFailed": "Swap transaction failed"
+      "swapTransactionFailed": "Swap transaction failed",
+      "insufficientBalance": "Insufficient balance. Maximum spendable: {{amount}} {{symbol}}"
     }
   },
   "swapProcessingScreen": {
